+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 📝 ***Note to self**, document the changes so that you can remember what you changed and why.*
 
+### 5/25/2025
+
+#### Added
+- Add **Microsoft.WinGet.Client**.
+
 ### 5/24/2025 - Init repo
 
 #### Added